Make the export / upload class for Template configurable
