+++ dradis-projects.gemspec
@@ -25,6 +25,6 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'rake', '~> 10.0'
   spec.add_development_dependency 'rspec'
 
-  spec.add_dependency 'dradis-plugins', '~> 3.5'
+  spec.add_dependency 'dradis-plugins', '~> 3.6'
   spec.add_dependency 'rubyzip', '~> 1.1.0'
 end
