+++ lib/dradis/plugins/projects/gem_version.rb
@@ -8,7 +8,7 @@ module Dradis
 
       module VERSION
         MAJOR = 3
-        MINOR = 0
+        MINOR = 1
         TINY = 0
         PRE = nil
